Small refactoring

+++ source/game_sa/Fx/FxEmitterPrt.cpp
@@ -4,6 +4,6 @@
 #include "FxManager.h"
 
 // 0x4A18B0
-void* FxEmitterPrt_c::operator new[](uint32 size) {
+void* FxEmitterPrt_c::operator new[](size_t size) {
     return g_fxMan.GetMemPool().GetMem(size);
 }
